cradle 동기화 처리 완료
