+++ Project_Spirit/Assets/Scripts/Event/RainDrop.cs
@@ -1,0 +1,35 @@
+ using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.UI;
+
+public class RainDrop : MonoBehaviour
+{
+    public Image image;
+    public Sprite[] sprites;
+   
+    public float frameRate = 0.1f;  // 프레임 속도 (초 단위)
+
+    private int currentFrame;
+    private float timer;
+
+    void Start()
+    {
+        if (sprites == null)
+        {
+            image = GetComponent<Image>();
+        }
+    }
+
+    void Update()
+    {
+        timer += Time.deltaTime;
+
+        if (timer >= frameRate)
+        {
+            timer -= frameRate;
+            currentFrame = (currentFrame + 1) % sprites.Length;
+            gameObject.GetComponent<Image>().sprite = sprites[currentFrame];
+        }
+    }
+}
