+++ Project_Spirit/Assets/Scripts/Event/RainDrop.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a9c3d6e38f445654485b589c80297fb6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
